move functions into auth module
