+++ corehq/apps/domain/decorators.py
@@ -2,7 +2,6 @@
 import base64
 from functools import wraps
 import logging
-import re
 
 # Django imports
 from django.conf import settings
@@ -16,6 +15,7 @@ from django.utils.translation import ugettext as _
 
 # External imports
 from django_digest.decorators import httpdigest
+from corehq.apps.domain.auth import determine_authtype_from_request
 from django_prbac.utils import has_privilege
 
 from django.http import HttpResponse
@@ -183,51 +183,6 @@ def login_or_basic_ex(allow_cc_users=False):
     return _login_or_challenge(basicauth(), allow_cc_users=allow_cc_users)
 
 login_or_basic = login_or_basic_ex()
-
-
-J2ME = 'j2me'
-ANDROID = 'android'
-
-
-def guess_phone_type_from_user_agent(user_agent):
-    """
-    A really dumb utility that guesses the phone type based on the user-agent header.
-    """
-    j2me_pattern = '[Nn]okia|NOKIA|CLDC|cldc|MIDP|midp|Series60|Series40|[Ss]ymbian|SymbOS|[Mm]aemo'
-    if user_agent:
-        if re.search(j2me_pattern, user_agent):
-            return J2ME
-        elif 'Android' in user_agent:
-            return ANDROID
-    return None
-
-
-def determine_authtype_from_header(request, default=None):
-    """
-    Guess the auth type, based on request.
-    """
-    auth_header = (request.META.get('HTTP_AUTHORIZATION') or '').lower()
-    if auth_header.startswith('basic '):
-        return 'basic'
-    elif auth_header.startswith('digest '):
-        return 'digest'
-    elif all(ApiKeyAuthentication().extract_credentials(request)):
-        return 'api_key'
-
-    return default
-
-
-def determine_authtype_from_request(request, default='basic'):
-    user_agent = request.META.get('HTTP_USER_AGENT')
-    type_to_auth_map = {
-        J2ME: 'digest',
-        ANDROID: 'basic',
-    }
-    user_type = guess_phone_type_from_user_agent(user_agent)
-    if user_type is not None:
-        return type_to_auth_map.get(user_type, default)
-    else:
-        return determine_authtype_from_header(request, default=default)
 
 
 def login_or_digest_or_basic(default='basic'):
